adding index.html to templates and using template to serve page also added functions to handle js requests. updated readme
note:
handling the requests for files in the js folder needs to be fixed.
for example, when a request for a url such as /shop/1 is used, it will
look in /shop/1/js/... to get the needed files.  this somehow needs to be
redirected to the proper /static/js/ location